abs() takes an int so use atoi() to match types.
git-svn-id: svn://svn.code.sf.net/p/fuse-emulator/code/trunk@5265 82cbdd72-22ac-47d8-a7df-4a7c34dd7743

+++ fuse-utils/tape2wav.c
@@ -66,7 +66,7 @@ main( int argc, char **argv )
 
     switch( c ) {
 
-    case 'r': sample_rate = abs( atol( optarg ) ); break;
+    case 'r': sample_rate = abs( atoi( optarg ) ); break;
     case 'h': show_help(); return 0;
     case 'V': show_version(); return 0;
